camera/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
pillow
python-multipart
onnx 
google-cloud-storage